backend/requirements.txt
# =====================================================================
# requirements.txt - Complete Python Dependencies
# =====================================================================

flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.28
crewai==0.41.1
crewai-tools==0.4.26
agentops==0.2.7
pysbd==0.3.4
Werkzeug==3.0.1
click>=8.0
itsdangerous>=2.0
Jinja2>=3.0
MarkupSafe>=2.0
six>=1.9.0
urllib3>=1.26.0
certifi>=2021.5.25
charset-normalizer>=2.0.0
idna>=2.5
python-dateutil>=2.7.0
pytz>=2020.1
lxml>=4.6.0
beautifulsoup4>=4.9.0
html5lib>=1.1
multitasking>=0.0.7
frozendict>=2.3.4
peewee>=3.16.2
appdirs>=1.4.4
